feature: implemented update (PUT) methods

+++ src/main/java/com/SunnyGadgetsProject/SunnyGadgets_v1/service/ServicePermission.java
@@ -11,10 +11,11 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 
 @Service
-public class ServicePermission  implements  IServicePermission{
+public class ServicePermission implements IServicePermission {
 
     private final IRepositoryPermission permissionRepository;
     private final PermissionMapper permissionMapper;
@@ -34,12 +35,14 @@ public class ServicePermission  implements  IServicePermission{
         if (permissionResponseDTOS.isEmpty()) {
             throw new EntityNotFoundException("No permission's found"); //Exception Not Found
         }
+        logger.info("Get all permissions from repository");
         return permissionResponseDTOS;
     }
 
     @Override
     public PermissionResponseDTO getPermissionById(Long id) {
         Permission permission = permissionRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+        logger.info("Get permission by id: " + id);
         return permissionMapper.toDto(permission);
     }
 
@@ -55,26 +58,38 @@ public class ServicePermission  implements  IServicePermission{
     @Override
     public List<PermissionResponseDTO> createPermission(Set<PermissionCreateDTO> permissions) {
         List<Permission> per = new ArrayList<>();
-        List<PermissionResponseDTO> customerDTOList = new ArrayList<>();
+        List<PermissionResponseDTO> permissionDTOList = new ArrayList<>();
         for (PermissionCreateDTO pe : permissions) {
             per.add(permissionMapper.toEntity(pe));
         }
         permissionRepository.saveAll(per);
         for (Permission permission : per) {
-            customerDTOList.add(permissionMapper.toDto(permission));
+            permissionDTOList.add(permissionMapper.toDto(permission));
         }
-        logger.info("Customer's created: {}", per);
-        return customerDTOList;
+        logger.info("Permission's created: {}", per);
+        return permissionDTOList;
     }
 
 
     @Override
-    public Permission updatePermission(Permission permission) {
-        return permissionRepository.save(permission);
+    public PermissionResponseDTO updatePermission(PermissionCreateDTO permission, Long id) {
+        Optional<Permission> permissionOptional = permissionRepository.findById(id);
+        if (permissionOptional.isEmpty()) {
+            throw new EntityNotFoundException("Permission with id " + id + " not found"); //Exception Not Found
+        }
+
+        permissionOptional.get().setPermissionName(permission.getPermissionName());
+        permissionRepository.save(permissionOptional.get());
+        logger.info("Permission updated: {}", permission.getPermissionName());
+        return permissionMapper.toDto(permissionOptional.get());
     }
 
     @Override
     public void deletePermission(Long id) {
+        if (permissionRepository.findById(id).isEmpty()) {
+            throw new EntityNotFoundException("Permission with id " + id + " not found");
+        }
+        logger.info("Permission deleted: {}", id);
         permissionRepository.deleteById(id);
     }
 }
